Rename folder auth -> session
